CU-86c29qw7c - Refine Webserv and Server (#12)

+++ srcs/Server.cpp
@@ -2,8 +2,13 @@
 
 Server::Server() : _ctx("./var/www"), _ctx_err("./var/error")
 {
+  Location* default_location = new Location;
+  default_location->set("key", "/");
+  _locations["/"] = default_location;
+  
   _setters["listen"] = &Server::setListen;
   _setters["server_name"] = &Server::setServerName;
+
   _setters["root"] = &Server::setRoot;
   _setters["index"] = &Server::setIndex;
   _setters["error_page"] = &Server::setErrorPage;
@@ -14,6 +19,7 @@ Server::Server() : _ctx("./var/www"), _ctx_err("./var/error")
   _getters["server_name"] = &Server::getServerName;
   _getters["listen"] = &Server::getListen;
   _getters["index"] = &Server::getIndex;
+
 }
 
 Server::Server(const Server& cpy)
@@ -34,7 +40,6 @@ Server::~Server()
     it->second = NULL;
   }
 }
-
 
 std::ostream& operator<<(std::ostream& cout, const Server& server)
 {
@@ -68,6 +73,8 @@ bool  Server::set(const std::string key, Location* location)
 
 std::string  Server::get(const std::string key) const 
 {
+  if (this->_getters.find(key) == this->_getters.end())
+    return "";
   return (this->*_getters.at(key))();
 }
 
@@ -75,14 +82,6 @@ void  Server::setDefault()
 {
   if (this->_listen.empty())
     this->set("listen", "80");
-  if (this->_index.empty())
-    this->_index = "index.html";
-  if (this->_error_page.empty())
-    this->_error_page = "error.html";
-  if (this->_allow_listing.empty())
-    this->_allow_listing = "false";
-  if (this->_allowed_method.empty())
-    this->_allowed_method = "GET/POST/DELETE";
 }
 
 std::vector<std::string>::iterator  Server::iterator(int pos)
@@ -103,7 +102,7 @@ void  Server::handle(Request& request)
   
   Location* location = this->getLocation(ressource);
   
-  this->handlePath(ressource, location, request);
+  this->handlePath(ressource, request, location);
 
   std::string ext = Parser::getExtension(ressource);
   std::string body;
@@ -113,14 +112,11 @@ void  Server::handle(Request& request)
   else 
     body = this->handleAction(ressource, request, location);
 
-  std::cout << "body : '" <<  body << "'" << std::endl;
-
-  // 3. Construire la response a partir du chemin + body 
   this->_response.generate(body, request);
   request.clean();
 }
 
-void  Server::handlePath(std::string& ressource, Location* location, Request& request)
+void  Server::handlePath(std::string& ressource,  Request& request, Location* location)
 {
   if (request.getStatus() == Request::E_REQUEST_BAD)
     return this->handleError(request.getResponseCode(), ressource, location);
@@ -139,7 +135,6 @@ void  Server::handlePath(std::string& ressource, Location* location, Request& re
 
 bool  Server::checkPathAccess(std::string method, std::string& ressource)
 {
-  std::cout << this->_response.getIsListing() << std::endl;
   if (!method.compare("GET") && access(ressource.c_str(), R_OK) == -1)
     return false;
 
@@ -154,9 +149,7 @@ bool  Server::checkPathAccess(std::string method, std::string& ressource)
 
 bool  Server::checkPathType(std::string method, std::string& ressource, Location* location)
 {
-  bool  allowed = this->isListingAllowed();
-  if (location && location->isset("allow_listing"))
-    allowed = location->isListingAllowed();
+  bool  allowed = location->isset("allow_listing");
   struct stat sb;
   if (stat(ressource.c_str(), &sb) == -1) 
     return false;
@@ -174,16 +167,16 @@ bool  Server::checkPathType(std::string method, std::string& ressource, Location
 
 void  Server::addPathRoot(std::string& ressource, Location* location)
 {
-  if (location)
+  std::string pattern = location->get("key");
+  if (pattern.compare("/"))
   {
-    std::string pattern = location->get("key");
     if (location->isset("root"))
       ressource.replace(0, pattern.length(), location->get("root"));
     else
-      ressource.replace(0, pattern.length(), this->_root);
+      ressource.replace(0, pattern.length(), this->get("root"));
   }
   else
-    ressource.insert(0, this->_root);
+    ressource.insert(0, location->get("root"));
   
   if (ressource[0] != '/')
     ressource.insert(0, "/");  
@@ -195,20 +188,20 @@ void  Server::addPathIndex(std::string& ressource, Location* location)
 {
   if (ressource[ressource.length() - 1] != '/')
     ressource.append("/");
-  if (location && location->isset("index")) 
+  if (location->isset("index")) 
     ressource.append(location->get("index"));
   else
-    ressource.append(this->_index);
+    ressource.append(this->get("index"));
 }
 
 Location* Server::getLocation(const std::string& ressource)
 {
   for (std::map<std::string, Location*>::iterator it = this->_locations.begin(); it != this->_locations.end(); ++it)
   {
-    if (ressource.find(it->first) == 0)
+    if (ressource.find(it->first) == 0 && it->first.compare("/") && (ressource.length() == it->first.length() || ressource[it->first.length() + 1] == '/'))
       return it->second;
   }
-  return NULL;
+  return this->_locations.at("/");
 }
 
 std::string Server::handleCgi()
@@ -223,22 +216,27 @@ std::string Server::handleCgi()
 std::string Server::handleAction(const std::string& ressource, Request& request, Location* location)
 {
   std::string method = request.getMethod();
-  // Verifier si la method est allow 
+  if (location->isset("redirect"))
+    return this->handleRedirect(location);
+  if (!location->isset("allowed_method", method))
+  {
+    std::string err = ressource;
+    this->handleError(AHttpMessage::NOT_ALLOWED, err, location);
+    return this->handleGet(err, location);
+  }
   if (!method.compare("GET"))
-    return this->handleGet(ressource);
+    return this->handleGet(ressource, location);
   if (!method.compare("DELETE"))
     return this->handleDelete(ressource, location);
   if (!method.compare("POST"))
-    return this->handlePost(ressource, request.getBody(), location);
-  std::string err = ressource;
-  this->handleError(AHttpMessage::NOT_ALLOWED, err, location);
-  return this->handleGet(err);
+    return this->handlePost(ressource, request, location);
+  
+  return "";
 }
 
-std::string Server::handleGet(const std::string& path)
+std::string Server::handleGet(const std::string& path, Location* location)
 {
-  // Verifier si le listing est allow 
-  if (this->isListingAllowed() && this->_response.getIsListing())
+  if (location->isset("allow_listing") && this->_response.getIsListing())
     return this->handleListing(path);
   std::ifstream ressource(path.c_str());
   if (!ressource)
@@ -259,22 +257,40 @@ std::string Server::handleDelete(const std::string& path, Location* location)
   {
     std::string ressource = path;
     this->handleError(this->_response.getResponseCode(), ressource, location);
-    return this->handleGet(ressource);
+    return this->handleGet(ressource, location);
   }
   return "";
 }
 
-std::string Server::handlePost(const std::string& path, const std::string body, Location* location)
+std::string Server::handlePost(const std::string& path, const Request& request, Location* location)
 {
-  std::ofstream file(path.c_str());
+  std::ofstream file;
+  if (request.getIsBinary())
+    file.open(path.c_str(), std::ios::binary);
+  else
+    file.open(path.c_str());
+
   if (!file || this->_response.getResponseCode().compare("200 OK"))
   {
     std::string ressource = path;
     this->handleError(this->_response.getResponseCode(), ressource, location);
-    return this->handleGet(ressource);
+    return this->handleGet(ressource, location);
   }  
-  file << body;
+  file << request.getBody();
   file.close();
+  return "";
+}
+
+std::string  Server::handleRedirect(Location* location)
+{
+  std::string code = this->_response.get(location->get("redirect", "code"));
+
+  this->_response.setResponseCode(code);
+
+  this->_response.setIsRedirect(true);
+
+  this->_response.setAttribute("Location", location->get("redirect", "url"));
+
   return "";
 }
 
@@ -304,32 +320,22 @@ std::string  Server::handleListing(const std::string& path)
 
 void  Server::handleError(std::string code, std::string& ressource, Location* location)
 {
-  std::cout << "Handle error : " << code << std::endl;
-  
   this->_response.setResponseCode(code);
 
-  std::string error_default = "error.html";
-
-  if (location && location->isset("error_page"))
-  {
-    std::string error_page = location->get("error_page");
-    if (access(error_page.c_str(), R_OK) == 0)
-    {
-      ressource = error_page;
-      return ;
-    }
-  }
-
-  if (access(this->_error_page.c_str(), R_OK) == 0)
-    ressource = this->_error_page;
-  else
-    ressource = error_default;
+  if (location->isset("error_page"))
+    ressource = location->get("error_page", code);
 
   if (ressource[0] != '/')
     ressource.insert(0, "/");  
-  
+
   ressource.insert(0, this->_ctx_err);
-  return ;
+  
+  struct stat sb;
+  if (stat(ressource.c_str(), &sb) == -1 || S_ISDIR(sb.st_mode) || access(ressource.c_str(), R_OK) == -1) 
+  {
+    ressource = this->_ctx_err.substr(0);
+    ressource.append("/error.html");
+  }
 }
 
 bool  Server::send(int fd)
@@ -352,14 +358,10 @@ std::vector<std::string>::const_iterator  Server::iterator(int pos) const
   return it;
 }
 
-bool  Server::isListingAllowed() const 
-{
-  return !this->_allow_listing.compare("true") || !this->_allow_listing.compare("TRUE");
-}
-
-
 bool  Server::isset(const std::string key) const
 {
+  if (this->_getters.find(key) == this->_getters.end())
+    return false;
   return (!(this->*_getters.at(key))().empty());
 }
 
@@ -379,40 +381,38 @@ void  Server::setServerName(std::string& value)
 
 void  Server::setRoot(std::string& value)
 {
-  if (!this->_root.empty())
-    Validator::throwError("Try to rewrite root directory.");
-  this->_root = value;
+  this->_locations.at("/")->set("root", value);
 }
 
 void  Server::setIndex(std::string& value)
 {
   if (!Validator::validate("file", value))
     Validator::throwError("Invalid file for index.");
-  this->_index = value;
+  this->_locations.at("/")->set("index", value);
 }
 
 void  Server::setErrorPage(std::string& value)
 {
   if (!Validator::validate("file", value))
     Validator::throwError("Error in error_page instruction.");
-  this->_error_page = value;
+  this->_locations.at("/")->set("error_page", value);
 }
 
 void  Server::setAllowedMethod(std::string& value)
 {
   if (!Validator::validate("method", value))
     Validator::throwError("Error in allowed_method instruction.");
-  this->_allowed_method = value;
+  this->_locations.at("/")->set("allowed_method", value);
 }
 
 void  Server::setAllowListing(std::string& value)
 {
-  this->_allow_listing = value;
+  this->_locations.at("/")->set("allow_listing", value);
 }
 
 void  Server::setRedirect(std::string& value)
 {
-  this->_redirect = value;
+  this->_locations.at("/")->set("redirect", value);
 }
 
 std::string  Server::getServerName() const
@@ -427,5 +427,30 @@ std::string  Server::getListen() const
 
 std::string  Server::getIndex() const
 {
-  return this->_index;
+  return this->_locations.at("/")->get("index");
+}
+
+std::string  Server::getRoot() const
+{
+  return this->_locations.at("/")->get("root");
+}
+
+std::string  Server::getAllowedMethod() const
+{
+  return this->_locations.at("/")->get("allowed_method");
+}
+
+std::string  Server::getAllowListing() const
+{
+  return this->_locations.at("/")->get("allow_listing");
+}
+
+std::string  Server::getRedirect() const
+{
+  return this->_locations.at("/")->get("redirect");
+}
+
+std::string  Server::getErrorPage() const
+{
+  return this->_locations.at("/")->get("error_page");
 }
